Создал admin_commands.py, в нём admin_commands_router, также в нём добавил возможности: получить кол-во пользователей, информацию по конкретному пользователю, отправку сообщений всем пользователям. Сделал сумму на демо-счёте вещественным, а не целым числом. Исправил добавление нового пользователя в basic_commands.py, также в этом файле добавил команду "/admin". Создал отдельные клавиатуры и обработчики для всех вариантов команды "/update_demo_account". Сделал подтверждение для отправки сообщения всем пользователям. Добавил сообщения ожидания в get_information_commands.py. В Bot.py добавил список id админов; добавил функцию для удаления сообщения (с кнопками, если они не были нажаты). Сделал удаление сообщения с кнопками после нажатия или долгого ожидания для некоторых инлайнов. Исправил некоторые недоработки.

+++ Trading_bot_aiogram/on_start_update_data_base.py
@@ -43,8 +43,8 @@ async def on_start_update_data_base():
         CREATE TABLE IF NOT EXISTS users_demo_account (
                 Id INTEGER PRIMARY KEY,
                 is_demo_account TEXT,
-                start_sum INTEGER,
-                current_sum INTEGER,
+                start_sum REAL,
+                current_sum REAL,
                 is_auto_operation TEXT,
                 operation_percent INTEGER
     )
